poprawki

+++ compression/avar_gpu.cu
@@ -20,45 +20,45 @@ __device__ __inline__ unsigned int GETNPBITS(
 #define WARP_SIZE 32
 #define WORD_SIZE 32
 
-__global__ void avar_compress_gpu (avar_header comp_h, int *data, int *compressed_data, size_t length)
+__global__ void avar_compress_gpu (avar_header comp_h, int *data, int *compressed_data, unsigned long length)
 {
-    long warp_th = (threadIdx.x % WARP_SIZE); 
-    size_t pos = blockIdx.x * blockDim.x + threadIdx.x - warp_th;
-    size_t data_id = pos * WARP_SIZE + warp_th;
-    size_t cdata_id = pos * comp_h.bit_length + warp_th;
+    int warp_th = (threadIdx.x % WARP_SIZE); 
+    unsigned long pos = blockIdx.x * blockDim.x + threadIdx.x - warp_th;
+    unsigned long data_id = pos * WARP_SIZE + warp_th;
+    unsigned long cdata_id = pos * comp_h.bit_length + warp_th;
     avar_compress_base_gpu(comp_h, data_id, cdata_id, data, compressed_data, length);
 }
 
-__global__ void avar_decompress_gpu (avar_header comp_h, int *compressed_data, int * decompress_data, size_t length)
+__global__ void avar_decompress_gpu (avar_header comp_h, int *compressed_data, int * decompress_data, unsigned long length)
 {
     int warp_th = (threadIdx.x % WARP_SIZE); 
-    size_t pos = blockIdx.x * blockDim.x + threadIdx.x - warp_th;
-    size_t data_id = pos * WARP_SIZE + warp_th;
-    size_t cdata_id = pos * comp_h.bit_length + warp_th;
+    unsigned long pos = blockIdx.x * blockDim.x + threadIdx.x - warp_th;
+    unsigned long data_id = pos * WARP_SIZE + warp_th;
+    unsigned long cdata_id = pos * comp_h.bit_length + warp_th;
     avar_decompress_base_gpu(comp_h, cdata_id, data_id, compressed_data, decompress_data, length);
 }
 
 
-__host__ void run_avar_compress_gpu(avar_header comp_h, int *data, int *compressed_data, size_t length)
+__host__ void run_avar_compress_gpu(avar_header comp_h, int *data, int *compressed_data, unsigned long length)
 {
     int block_size = WARP_SIZE * 8; // better occupancy 
-    size_t block_number = (length + block_size * WARP_SIZE - 1) / (block_size * WARP_SIZE);
+    unsigned long block_number = (length + block_size * WARP_SIZE - 1) / (block_size * WARP_SIZE);
     avar_compress_gpu <<<block_number, block_size>>> (comp_h, data, compressed_data, length);
 }
 
-__host__ void run_avar_decompress_gpu(avar_header comp_h, int *compressed_data, int *data, size_t length)
+__host__ void run_avar_decompress_gpu(avar_header comp_h, int *compressed_data, int *data, unsigned long length)
 {
     int block_size = WARP_SIZE * 8; // better occupancy
-    size_t block_number = (length + block_size * WARP_SIZE - 1) / (block_size * WARP_SIZE);
+    unsigned long block_number = (length + block_size * WARP_SIZE - 1) / (block_size * WARP_SIZE);
     avar_decompress_gpu <<<block_number, block_size>>> (comp_h, compressed_data, data, length);
 }
 
 
-__device__  void avar_compress_base_gpu (avar_header comp_h, size_t data_id, size_t comp_data_id, int *data, int *compressed_data, size_t length)
+__device__  void avar_compress_base_gpu (avar_header comp_h, unsigned long data_id, unsigned long comp_data_id, int *data, int *compressed_data, unsigned long length)
 {
     int v1, value = 0;
     unsigned int v1_pos=0, v1_len;
-    size_t pos=comp_data_id, pos_decomp=data_id;
+    unsigned long pos=comp_data_id, pos_decomp=data_id;
 
     for (unsigned int i = 0; i < WARP_SIZE && pos_decomp < length; ++i)
     {
@@ -87,9 +87,9 @@ __device__  void avar_compress_base_gpu (avar_header comp_h, size_t data_id, siz
     }
 }
 
-__device__ void avar_decompress_base_gpu (avar_header comp_h, size_t comp_data_id, size_t data_id, int *compressed_data, int *data, size_t length)
+__device__ void avar_decompress_base_gpu (avar_header comp_h, unsigned long comp_data_id, unsigned long data_id, int *compressed_data, int *data, unsigned long length)
 {
-    size_t pos=comp_data_id, pos_decomp=data_id;
+    unsigned long pos=comp_data_id, pos_decomp=data_id;
     unsigned int v1_pos=0, v1_len;
     int v1, ret;
